break timestamp stuff into its own package
